# First draft of normalized exposure rate