@DavidKerkmann 2025-07-31
Performance benchmarks:

Main:
```
----------------------------------------------------------------------------------
Benchmark                                        Time             CPU   Iterations   
----------------------------------------------------------------------------------   
abm_benchmark/abm_benchmark_50k               3700 ms         3688 ms            1   
abm_benchmark/abm_benchmark_50k               3410 ms         3406 ms            1   
abm_benchmark/abm_benchmark_50k               3361 ms         3359 ms            1   
abm_benchmark/abm_benchmark_50k               3435 ms         3422 ms            1   
abm_benchmark/abm_benchmark_50k               3311 ms         3312 ms            1   
abm_benchmark/abm_benchmark_50k_mean          3443 ms         3438 ms            5   
abm_benchmark/abm_benchmark_50k_median        3410 ms         3406 ms            5   
abm_benchmark/abm_benchmark_50k_stddev         151 ms          146 ms            5   
abm_benchmark/abm_benchmark_50k_cv            4.40 %          4.25 %             5
abm_benchmark/abm_benchmark_100k              6743 ms         6734 ms            1
abm_benchmark/abm_benchmark_100k              6996 ms         6938 ms            1   
abm_benchmark/abm_benchmark_100k              6813 ms         6797 ms            1   
abm_benchmark/abm_benchmark_100k              6664 ms         6641 ms            1   
abm_benchmark/abm_benchmark_100k              7303 ms         7312 ms            1   
abm_benchmark/abm_benchmark_100k_mean         6904 ms         6884 ms            5   
abm_benchmark/abm_benchmark_100k_median       6813 ms         6797 ms            5   
abm_benchmark/abm_benchmark_100k_stddev        255 ms          263 ms            5   
abm_benchmark/abm_benchmark_100k_cv           3.69 %          3.81 %             5   
abm_benchmark/abm_benchmark_200k             14068 ms        13844 ms            1
abm_benchmark/abm_benchmark_200k             13513 ms        13484 ms            1   
abm_benchmark/abm_benchmark_200k             13415 ms        13391 ms            1   
abm_benchmark/abm_benchmark_200k             12679 ms        12688 ms            1   
abm_benchmark/abm_benchmark_200k             13039 ms        13000 ms            1   
abm_benchmark/abm_benchmark_200k_mean        13343 ms        13281 ms            5   
abm_benchmark/abm_benchmark_200k_median      13415 ms        13391 ms            5   
abm_benchmark/abm_benchmark_200k_stddev        523 ms          448 ms            5   
abm_benchmark/abm_benchmark_200k_cv           3.92 %          3.37 %             5 
```


This branch:
```
----------------------------------------------------------------------------------
Benchmark                                        Time             CPU   Iterations
----------------------------------------------------------------------------------
abm_benchmark/abm_benchmark_50k               3566 ms         3562 ms            1
abm_benchmark/abm_benchmark_50k               3463 ms         3453 ms            1
abm_benchmark/abm_benchmark_50k               3527 ms         3531 ms            1
abm_benchmark/abm_benchmark_50k               3585 ms         3578 ms            1
abm_benchmark/abm_benchmark_50k               3365 ms         3359 ms            1
abm_benchmark/abm_benchmark_50k_mean          3501 ms         3497 ms            5
abm_benchmark/abm_benchmark_50k_median        3527 ms         3531 ms            5
abm_benchmark/abm_benchmark_50k_stddev        89.1 ms         90.7 ms            5
abm_benchmark/abm_benchmark_50k_cv            2.55 %          2.59 %             5
abm_benchmark/abm_benchmark_100k              7511 ms         7500 ms            1
abm_benchmark/abm_benchmark_100k              7151 ms         7141 ms            1
abm_benchmark/abm_benchmark_100k              6821 ms         6812 ms            1
abm_benchmark/abm_benchmark_100k              6882 ms         6891 ms            1
abm_benchmark/abm_benchmark_100k              6894 ms         6891 ms            1
abm_benchmark/abm_benchmark_100k_mean         7052 ms         7047 ms            5
abm_benchmark/abm_benchmark_100k_median       6894 ms         6891 ms            5
abm_benchmark/abm_benchmark_100k_stddev        286 ms          282 ms            5
abm_benchmark/abm_benchmark_100k_cv           4.06 %          4.00 %             5
abm_benchmark/abm_benchmark_200k             13002 ms        13000 ms            1
abm_benchmark/abm_benchmark_200k             12257 ms        12250 ms            1
abm_benchmark/abm_benchmark_200k             12183 ms        12172 ms            1
abm_benchmark/abm_benchmark_200k             12214 ms        12219 ms            1
abm_benchmark/abm_benchmark_200k             12333 ms        12328 ms            1
abm_benchmark/abm_benchmark_200k_mean        12398 ms        12394 ms            5
abm_benchmark/abm_benchmark_200k_median      12257 ms        12250 ms            5
abm_benchmark/abm_benchmark_200k_stddev        343 ms          344 ms            5
abm_benchmark/abm_benchmark_200k_cv           2.76 %          2.77 %             5
```